DEV: Move form data handling to separate `lib`
